+++ README.md
@@ -1,0 +1,6 @@
+# Simple Web Framework
+
+
+## build project
+
+$ mvn clean install
