+++ src/main/java/com/pollra/client/controller/DataBaseController.java
@@ -1,4 +1,4 @@
-package main.java.pollra.client.controller;
+package com.pollra.client.controller;
 
 /**
  *
