+++ src/main/java/com/pollra/client/controller/ResponseController.java
@@ -1,14 +1,11 @@
-package main.java.pollra.client.controller;
+package com.pollra.client.controller;
 
 
-import main.java.pollra.client.Resolver.LoginResolver;
-import main.java.pollra.client.Resolver.ViewResolver;
-import main.java.pollra.client.http.HttpHeader;
-import main.java.pollra.client.http.HttpRequest;
-import main.java.pollra.client.http.HttpResponse;
-import main.java.pollra.client.http.HttpStatus;
+import com.pollra.client.Resolver.LoginResolver;
+import com.pollra.client.http.HttpHeader;
+import com.pollra.client.http.HttpRequest;
+import com.pollra.client.http.HttpStatus;
 
-import java.util.HashMap;
 import java.util.Map;
 
 /**
