+++ src/main/java/com/pollra/client/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package main.java.pollra.client.http;
+package com.pollra.client.http;
 
 public enum HttpMethod {
     GET, POST
