+++ src/main/java/com/pollra/client/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package main.java.pollra.client.http;
+package com.pollra.client.http;
 
 import java.util.Arrays;
 import java.util.HashMap;
