--- src/main/resources/README.md
@@ -1,1 +1,0 @@
-
